+++ lib/internal/bootstrap/pre_execution.js
@@ -67,7 +67,7 @@ function setupCoverageHooks(dir) {
   const {
     writeCoverage,
     setCoverageDirectory
-  } = require('internal/coverage-gen/with_profiler');
+  } = require('internal/profiler');
   setCoverageDirectory(coverageDirectory);
   process.on('exit', writeCoverage);
   process.reallyExit = (code) => {
