+++ lib/internal/process/per_thread.js
@@ -158,7 +158,7 @@ function wrapProcessMethods(binding) {
   function kill(pid, sig) {
     var err;
     if (process.env.NODE_V8_COVERAGE) {
-      const { writeCoverage } = require('internal/coverage-gen/with_profiler');
+      const { writeCoverage } = require('internal/profiler');
       writeCoverage();
     }
