+++ lib/internal/profiler.js
@@ -21,23 +21,16 @@ function writeCoverage() {
   }
 
   const target = join(coverageDirectory, filename);
-  try {
-    disableAllAsyncHooks();
-    internalBinding('coverage').end((msg) => {
+  internalBinding('profiler').endCoverage((msg) => {
+    try {
       const coverageInfo = JSON.parse(msg).result;
       if (coverageInfo) {
         writeFileSync(target, JSON.stringify(coverageInfo));
       }
-    });
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-function disableAllAsyncHooks() {
-  const { getHookArrays } = require('internal/async_hooks');
-  const [hooks_array] = getHookArrays();
-  hooks_array.forEach((hook) => { hook.disable(); });
+    } catch (err) {
+      console.error(err);
+    }
+  });
 }
 
 function setCoverageDirectory(dir) {
